Added important files, including system GUI, Arduino, and Raspberry Pi code
The "arduino" folder contains the .ino files for sending and recieving data from the RPi.

The "rpi" folder contains the .py file that recieves and sends commands to the Arduino.

The "images" folder contains the .PNG files required to run the GUI.

The "vsc" folder contains the GUI code. Keep in mind the image paths are absolute, and not relative.